доработана работа с картой
