+++ project/src/components/places/places.tsx
@@ -8,55 +8,46 @@ type PlacesProps = {
   onCardFocusChange: (id: number | null) => void;
 };
 
-export const Places: FC<PlacesProps> = ({ places, onCardFocusChange }) => {
-
-  const handleHotelCardMouseLeave = () => onCardFocusChange(null);
-
-  const handleHotelCardMouseEnter = (hotelCard : number | null) => {
-    onCardFocusChange(hotelCard);
-  };
-
-  return (
-    <section className="cities__places places">
-      <h2 className="visually-hidden">Places</h2>
-      <b className="places__found">312 places to stay in Amsterdam</b>
-      <form className="places__sorting" action="#" method="get">
-        <span className="places__sorting-caption">Sort by</span>
-        <span className="places__sorting-type" tabIndex={0}>
+export const Places: FC<PlacesProps> = ({ places, onCardFocusChange }) => (
+  <section className="cities__places places">
+    <h2 className="visually-hidden">Places</h2>
+    <b className="places__found">312 places to stay in Amsterdam</b>
+    <form className="places__sorting" action="#" method="get">
+      <span className="places__sorting-caption">Sort by</span>
+      <span className="places__sorting-type" tabIndex={0}>
+        Popular
+        <svg className="places__sorting-arrow" width="7" height="4">
+          <use xlinkHref="#icon-arrow-select"></use>
+        </svg>
+      </span>
+      <ul className="places__options places__options--custom places__options--opened">
+        <li className="places__option places__option--active" tabIndex={0}>
           Popular
-          <svg className="places__sorting-arrow" width="7" height="4">
-            <use xlinkHref="#icon-arrow-select"></use>
-          </svg>
-        </span>
-        <ul className="places__options places__options--custom places__options--opened">
-          <li className="places__option places__option--active" tabIndex={0}>
-            Popular
-          </li>
-          <li className="places__option" tabIndex={0}>
-            Price: low to high
-          </li>
-          <li className="places__option" tabIndex={0}>
-            Price: high to low
-          </li>
-          <li className="places__option" tabIndex={0}>
-            Top rated first
-          </li>
-        </ul>
-      </form>
-      <div className="cities__places-list places__list tabs__content">
-        {places.map((place: Hotel) => (
-          <PlaceCard
-            onMouseOver={handleHotelCardMouseEnter}
-            onMouseLeave={handleHotelCardMouseLeave}
-            place={place}
-            key={place.id}
-            htmlClasses={{
-              imageWrapperClass: 'cities__image-wrapper',
-              rootPlaceCardClass: 'cities__card',
-            }}
-          />
-        ))}
-      </div>
-    </section>
-  );
-};
+        </li>
+        <li className="places__option" tabIndex={0}>
+          Price: low to high
+        </li>
+        <li className="places__option" tabIndex={0}>
+          Price: high to low
+        </li>
+        <li className="places__option" tabIndex={0}>
+          Top rated first
+        </li>
+      </ul>
+    </form>
+    <div className="cities__places-list places__list tabs__content">
+      {places.map((place: Hotel) => (
+        <PlaceCard
+          onMouseOver={onCardFocusChange}
+          onMouseLeave={onCardFocusChange}
+          place={place}
+          key={place.id}
+          htmlClasses={{
+            imageWrapperClass: 'cities__image-wrapper',
+            rootPlaceCardClass: 'cities__card',
+          }}
+        />
+      ))}
+    </div>
+  </section>
+);
