+++ project/src/types/city.ts
@@ -1,8 +1,10 @@
+import { Cities } from '../const';
+
 export type City = {
   location: {
     latitude: number,
     longitude: number,
     zoom: number,
   },
-  name: string,
+  name: Cities,
 }
