+++ project/src/utils.ts
@@ -1,10 +1,2 @@
 export const getRating = (rating : number) : string => `${rating / 5 * 100}%`;
 
-export enum Cities {
-  Paris = 'Paris',
-  Cologne = 'Cologne',
-  Brussels = 'Brussels',
-  Amsterdam = 'Amsterdam',
-  Hamburg = 'Hamburg',
-  Dusseldorf = 'Dusseldorf',
-}
